4.2.0

2920a97877e12e055c1333079b8d19cee7f33826

+++ migration-engine/connectors/sql-migration-connector/src/sql_renderer/mssql_renderer.rs
@@ -10,6 +10,7 @@ use datamodel::dml::PrismaValue;
 use indoc::{formatdoc, indoc};
 use native_types::{MsSqlType, MsSqlTypeParameter};
 use sql_schema_describer::{
+    mssql::MssqlSchemaExt,
     walkers::{
         ColumnWalker, EnumWalker, ForeignKeyWalker, IndexWalker, TableWalker, UserDefinedTypeWalker, ViewWalker,
     },
@@ -40,7 +41,7 @@ impl MssqlFlavour {
         }
     }
 
-    fn render_column(&self, column: &ColumnWalker<'_>) -> String {
+    fn render_column(&self, column: ColumnWalker<'_>) -> String {
         let column_name = self.quote(column.name());
 
         let r#type = render_column_type(column);
@@ -71,11 +72,10 @@ impl MssqlFlavour {
         format!("{} {}{}{}", column_name, r#type, nullability, default)
     }
 
-    fn render_references(&self, foreign_key: &ForeignKeyWalker<'_>) -> String {
+    fn render_references(&self, foreign_key: ForeignKeyWalker<'_>) -> String {
         let cols = foreign_key
-            .referenced_column_names()
-            .iter()
-            .map(Quoted::mssql_ident)
+            .referenced_columns()
+            .map(|c| Quoted::mssql_ident(c.name()))
             .join(",");
 
         format!(
@@ -93,51 +93,42 @@ impl SqlRenderer for MssqlFlavour {
         Quoted::mssql_ident(name)
     }
 
-    fn render_alter_table(&self, alter_table: &AlterTable, schemas: &Pair<&SqlSchema>) -> Vec<String> {
-        let AlterTable {
-            table_ids: table_index,
-            changes,
-        } = alter_table;
-        let tables = schemas.tables(table_index);
-
+    fn render_alter_table(&self, alter_table: &AlterTable, schemas: Pair<&SqlSchema>) -> Vec<String> {
+        let AlterTable { table_ids, changes } = alter_table;
+        let tables = schemas.walk(*table_ids);
         alter_table::create_statements(self, tables, changes)
     }
 
-    fn render_alter_enum(&self, _: &AlterEnum, _: &Pair<&SqlSchema>) -> Vec<String> {
+    fn render_alter_enum(&self, _: &AlterEnum, _: Pair<&SqlSchema>) -> Vec<String> {
         unreachable!("render_alter_enum on Microsoft SQL Server")
     }
 
-    fn render_rename_index(&self, indexes: Pair<&IndexWalker<'_>>) -> Vec<String> {
+    fn render_rename_index(&self, indexes: Pair<IndexWalker<'_>>) -> Vec<String> {
         let index_with_table = format!(
             "{}.{}.{}",
             self.schema_name(),
-            indexes.previous().table().name(),
-            indexes.previous().name()
+            indexes.previous.table().name(),
+            indexes.previous.name()
         );
 
         vec![format!(
             "EXEC SP_RENAME N'{index_with_table}', N'{index_new_name}', N'INDEX'",
             index_with_table = index_with_table,
-            index_new_name = indexes.next().name(),
+            index_new_name = indexes.next.name(),
         )]
     }
 
-    fn render_create_enum(&self, _: &EnumWalker<'_>) -> Vec<String> {
+    fn render_create_enum(&self, _: EnumWalker<'_>) -> Vec<String> {
         unreachable!("render_create_enum on Microsoft SQL Server")
     }
 
-    fn render_create_index(&self, index: &IndexWalker<'_>) -> String {
-        let index_type = match index.index_type() {
-            IndexType::Unique => "UNIQUE ",
-            IndexType::Normal => "",
-            IndexType::Fulltext => unreachable!("Fulltext index on SQL Server"),
-        };
-
+    fn render_create_index(&self, index: IndexWalker<'_>) -> String {
+        let mssql_schema_ext: &MssqlSchemaExt = index.schema.downcast_connector_data();
         let index_name = self.quote(index.name());
         let table_reference = self.quote_with_schema(index.table().name()).to_string();
 
         let columns = index.columns().map(|c| {
-            let mut rendered = format!("{}", self.quote(c.get().name()));
+            let mut rendered = Quoted::mssql_ident(c.as_column().name()).to_string();
 
             if let Some(sort_order) = c.sort_order() {
                 rendered.push(' ');
@@ -147,29 +138,39 @@ impl SqlRenderer for MssqlFlavour {
             rendered
         });
 
-        format!(
-            "CREATE {index_type}INDEX {index_name} ON {table_reference}({columns})",
-            index_type = index_type,
-            index_name = index_name,
-            table_reference = table_reference,
-            columns = columns.join(", "),
-        )
+        let clustering = if mssql_schema_ext.index_is_clustered(index.id) {
+            "CLUSTERED "
+        } else {
+            "NONCLUSTERED "
+        };
+
+        let columns = columns.join(", ");
+
+        match index.index_type() {
+            IndexType::Unique => {
+                let constraint_name = self.quote(index.name());
+
+                format!("ALTER TABLE {table_reference} ADD CONSTRAINT {constraint_name} UNIQUE {clustering}({columns})")
+            }
+            IndexType::Normal => {
+                format!("CREATE {clustering}INDEX {index_name} ON {table_reference}({columns})",)
+            }
+            IndexType::Fulltext => todo!("SQL Server full-text indices..."),
+            IndexType::PrimaryKey => unreachable!(),
+        }
     }
 
-    fn render_create_table_as(&self, table: &TableWalker<'_>, table_name: &str) -> String {
-        let columns: String = table
-            .columns()
-            .map(|column| self.render_column(&column))
-            .join(",\n    ");
+    fn render_create_table_as(&self, table: TableWalker<'_>, table_name: &str) -> String {
+        let columns: String = table.columns().map(|column| self.render_column(column)).join(",\n    ");
+        let mssql_schema_ext: &MssqlSchemaExt = table.schema.downcast_connector_data();
 
         let primary_key = if let Some(pk) = table.primary_key() {
             let column_names = pk
-                .columns
-                .iter()
+                .columns()
                 .map(|col| {
-                    let mut rendered = format!("{}", self.quote(col.name()));
+                    let mut rendered = Quoted::mssql_ident(col.name()).to_string();
 
-                    if let Some(sort_order) = col.sort_order {
+                    if let Some(sort_order) = col.sort_order() {
                         rendered.push(' ');
                         rendered.push_str(sort_order.as_ref());
                     }
@@ -178,26 +179,27 @@ impl SqlRenderer for MssqlFlavour {
                 })
                 .join(",");
 
-            format!(
-                ",\n    CONSTRAINT {} PRIMARY KEY ({})",
-                self.quote(pk.constraint_name.as_ref().unwrap()),
-                column_names
-            )
+            let clustering = if mssql_schema_ext.index_is_clustered(pk.id) {
+                " CLUSTERED"
+            } else {
+                " NONCLUSTERED"
+            };
+
+            let constraint_name = Quoted::mssql_ident(pk.name());
+
+            format!(",\n    CONSTRAINT {constraint_name} PRIMARY KEY{clustering} ({column_names})",)
         } else {
             String::new()
         };
 
-        let constraints = table
-            .indexes()
-            .filter(|index| index.index_type().is_unique())
-            .collect::<Vec<_>>();
+        let constraints = table.indexes().filter(|index| index.is_unique()).collect::<Vec<_>>();
 
         let constraints = if !constraints.is_empty() {
             let constraints = constraints
                 .iter()
                 .map(|index| {
                     let columns = index.columns().map(|col| {
-                        let mut rendered = format!("{}", self.quote(col.get().name()));
+                        let mut rendered = format!("{}", self.quote(col.as_column().name()));
 
                         if let Some(sort_order) = col.sort_order() {
                             rendered.push(' ');
@@ -207,7 +209,16 @@ impl SqlRenderer for MssqlFlavour {
                         rendered
                     });
 
-                    format!("CONSTRAINT {} UNIQUE ({})", self.quote(index.name()), columns.join(","))
+                    let constraint_name = self.quote(index.name());
+                    let column_names = columns.join(",");
+
+                    let clustering = if mssql_schema_ext.index_is_clustered(index.id) {
+                        " CLUSTERED"
+                    } else {
+                        " NONCLUSTERED"
+                    };
+
+                    format!("CONSTRAINT {constraint_name} UNIQUE{clustering} ({column_names})")
                 })
                 .join(",\n    ");
 
@@ -228,11 +239,11 @@ impl SqlRenderer for MssqlFlavour {
         )
     }
 
-    fn render_drop_enum(&self, _: &EnumWalker<'_>) -> Vec<String> {
+    fn render_drop_enum(&self, _: EnumWalker<'_>) -> Vec<String> {
         unreachable!("render_drop_enum on MSSQL")
     }
 
-    fn render_drop_foreign_key(&self, foreign_key: &ForeignKeyWalker<'_>) -> String {
+    fn render_drop_foreign_key(&self, foreign_key: ForeignKeyWalker<'_>) -> String {
         format!(
             "ALTER TABLE {table} DROP CONSTRAINT {constraint_name}",
             table = self.quote_with_schema(foreign_key.table().name()),
@@ -240,48 +251,50 @@ impl SqlRenderer for MssqlFlavour {
         )
     }
 
-    fn render_drop_index(&self, index: &IndexWalker<'_>) -> String {
-        match index.index_type() {
-            IndexType::Normal => format!(
-                "DROP INDEX {} ON {}",
-                self.quote(index.name()),
-                self.quote_with_schema(index.table().name())
-            ),
-            IndexType::Unique => format!(
+    fn render_drop_index(&self, index: IndexWalker<'_>) -> String {
+        let ext: &MssqlSchemaExt = index.schema.downcast_connector_data();
+
+        if ext.index_is_a_constraint(index.id) {
+            format!(
                 "ALTER TABLE {} DROP CONSTRAINT {}",
                 self.quote_with_schema(index.table().name()),
                 self.quote(index.name()),
-            ),
-            IndexType::Fulltext => unreachable!("Fulltext index on SQL Server"),
+            )
+        } else {
+            format!(
+                "DROP INDEX {} ON {}",
+                self.quote(index.name()),
+                self.quote_with_schema(index.table().name())
+            )
         }
     }
 
-    fn render_redefine_tables(&self, tables: &[RedefineTable], schemas: &Pair<&SqlSchema>) -> Vec<String> {
+    fn render_redefine_tables(&self, tables: &[RedefineTable], schemas: Pair<&SqlSchema>) -> Vec<String> {
         // All needs to be inside a transaction.
         let mut result = vec!["BEGIN TRANSACTION".to_string()];
 
         for redefine_table in tables {
-            let tables = schemas.tables(&redefine_table.table_ids);
+            let tables = schemas.walk(redefine_table.table_ids);
             // This is a copy of our new modified table.
-            let temporary_table_name = format!("_prisma_new_{}", &tables.next().name());
+            let temporary_table_name = format!("_prisma_new_{}", &tables.next.name());
 
             // If any of the columns is an identity, we should know about it.
             let needs_autoincrement = redefine_table
                 .column_pairs
                 .iter()
-                .any(|(column_indexes, _, _)| tables.columns(column_indexes).next().is_autoincrement());
+                .any(|(column_indexes, _, _)| schemas.walk(*column_indexes).next.is_autoincrement());
 
             // Let's make the [columns] nicely rendered.
             let columns: Vec<_> = redefine_table
                 .column_pairs
                 .iter()
-                .map(|(column_indexes, _, _)| tables.columns(column_indexes).next().name())
+                .map(|(column_indexes, _, _)| schemas.walk(*column_indexes).next.name())
                 .map(|c| self.quote(c).to_string())
                 .collect();
 
             // Drop the indexes on the table.
-            for index in tables.previous().indexes() {
-                result.push(self.render_drop_index(&index));
+            for index in tables.previous.indexes().filter(|idx| !idx.is_primary_key()) {
+                result.push(self.render_drop_index(index));
             }
 
             // Remove all constraints from our original table. This will allow
@@ -300,10 +313,10 @@ impl SqlRenderer for MssqlFlavour {
                     AND OBJECT_NAME(PARENT_OBJECT_ID) = '{table}'
                     AND SCHEMA_NAME(SCHEMA_ID) = '{schema}'
                 EXEC sp_executesql @SQL
-            "#, table = tables.previous().name(), schema = self.schema_name()});
+            "#, table = tables.previous.name(), schema = self.schema_name()});
 
             // Create the new table.
-            result.push(self.render_create_table_as(tables.next(), &temporary_table_name));
+            result.push(self.render_create_table_as(tables.next, &temporary_table_name));
 
             // We cannot insert into autoincrement columns by default. If we
             // have `IDENTITY` in any of the columns, we'll allow inserting
@@ -320,7 +333,7 @@ impl SqlRenderer for MssqlFlavour {
                 IF EXISTS(SELECT * FROM {table})
                     EXEC('INSERT INTO {tmp_table} ({columns}) SELECT {columns} FROM {table} WITH (holdlock tablockx)')"#,
                                     columns = columns.join(","),
-                                    table = self.quote_with_schema(tables.previous().name()),
+                                    table = self.quote_with_schema(tables.previous.name()),
                                     tmp_table = self.quote_with_schema(&temporary_table_name),
             });
 
@@ -333,14 +346,14 @@ impl SqlRenderer for MssqlFlavour {
             }
 
             // Drop the old, now empty table.
-            result.extend(self.render_drop_table(tables.previous().name()));
+            result.extend(self.render_drop_table(tables.previous.name()));
 
             // Rename the temporary table with the name defined in the migration.
-            result.push(self.render_rename_table(&temporary_table_name, tables.next().name()));
+            result.push(self.render_rename_table(&temporary_table_name, tables.next.name()));
 
             // Recreate the indexes.
-            for index in tables.next().indexes().filter(|i| !i.index_type().is_unique()) {
-                result.push(self.render_create_index(&index));
+            for index in tables.next.indexes().filter(|i| !i.is_unique() && !i.is_primary_key()) {
+                result.push(self.render_create_index(index));
             }
         }
 
@@ -359,7 +372,7 @@ impl SqlRenderer for MssqlFlavour {
         )
     }
 
-    fn render_add_foreign_key(&self, foreign_key: &ForeignKeyWalker<'_>) -> String {
+    fn render_add_foreign_key(&self, foreign_key: ForeignKeyWalker<'_>) -> String {
         let mut add_constraint = String::with_capacity(120);
 
         write!(
@@ -376,7 +389,7 @@ impl SqlRenderer for MssqlFlavour {
                 add_constraint,
                 "CONSTRAINT [FK__{}__{}] ",
                 foreign_key.table().name(),
-                foreign_key.constrained_column_names().join("__"),
+                foreign_key.constrained_columns().map(|c| c.name()).join("__"),
             )
             .unwrap();
         }
@@ -385,9 +398,8 @@ impl SqlRenderer for MssqlFlavour {
             add_constraint,
             "FOREIGN KEY ({})",
             foreign_key
-                .constrained_column_names()
-                .iter()
-                .map(|col| self.quote(col))
+                .constrained_columns()
+                .map(|col| Quoted::mssql_ident(col.name()))
                 .join(", ")
         )
         .unwrap();
@@ -401,7 +413,7 @@ impl SqlRenderer for MssqlFlavour {
         vec![format!("DROP TABLE {}", self.quote_with_schema(table_name))]
     }
 
-    fn render_drop_view(&self, view: &ViewWalker<'_>) -> String {
+    fn render_drop_view(&self, view: ViewWalker<'_>) -> String {
         format!("DROP VIEW {}", self.quote_with_schema(view.name()))
     }
 
@@ -438,17 +450,17 @@ impl SqlRenderer for MssqlFlavour {
         Some(sql)
     }
 
-    fn render_rename_foreign_key(&self, fks: &Pair<ForeignKeyWalker<'_>>) -> String {
+    fn render_rename_foreign_key(&self, fks: Pair<ForeignKeyWalker<'_>>) -> String {
         format!(
             r#"EXEC sp_rename '{schema}.{previous}', '{next}', 'OBJECT'"#,
             schema = self.schema_name(),
-            previous = fks.previous().constraint_name().unwrap(),
-            next = fks.next().constraint_name().unwrap(),
+            previous = fks.previous.constraint_name().unwrap(),
+            next = fks.next.constraint_name().unwrap(),
         )
     }
 }
 
-fn render_column_type(column: &ColumnWalker<'_>) -> Cow<'static, str> {
+fn render_column_type(column: ColumnWalker<'_>) -> Cow<'static, str> {
     fn format_u32_arg(arg: Option<u32>) -> String {
         match arg {
             None => "".to_string(),
@@ -514,11 +526,16 @@ fn render_default(default: &DefaultValue) -> Cow<'_, str> {
         DefaultKind::Value(PrismaValue::String(val)) | DefaultKind::Value(PrismaValue::Enum(val)) => {
             Quoted::mssql_string(escape_string_literal(val)).to_string().into()
         }
-        DefaultKind::Value(PrismaValue::Bytes(b)) => format!("0x{}", format_hex(b)).into(),
+        DefaultKind::Value(PrismaValue::Bytes(b)) => {
+            let mut out = String::with_capacity(b.len() * 2 + 2);
+            out.push_str("0x");
+            format_hex(b, &mut out);
+            out.into()
+        }
         DefaultKind::Now => "CURRENT_TIMESTAMP".into(),
         DefaultKind::Value(PrismaValue::DateTime(val)) => Quoted::mssql_string(val).to_string().into(),
         DefaultKind::Value(PrismaValue::Boolean(val)) => Cow::from(if *val { "1" } else { "0" }),
         DefaultKind::Value(val) => val.to_string().into(),
-        DefaultKind::Sequence(_) => "".into(),
+        DefaultKind::Sequence(_) | DefaultKind::UniqueRowid => unreachable!(),
     }
 }
